+++ BrowserMatrix.m
@@ -31,16 +31,16 @@ static NSString *albumDragType = @"Osirix Album drag";
 	{
 		if( [theEvent modifierFlags] & NSShiftKeyMask )
 		{
-			int start = [[self cells] indexOfObject: [[self selectedCells] objectAtIndex: 0]];
-			int end = [[self cells] indexOfObject: [self cellAtRow:row column:column]];
+			NSInteger start = [[self cells] indexOfObject: [[self selectedCells] objectAtIndex: 0]];
+			NSInteger end = [[self cells] indexOfObject: [self cellAtRow:row column:column]];
 			
 			[self setSelectionFrom:start to:end anchor:start highlight: YES];
 			
 		}
 		else if( [theEvent modifierFlags] & NSCommandKeyMask )
 		{
-			int start = [[self cells] indexOfObject: [[self selectedCells] objectAtIndex: 0]];
-			int end = [[self cells] indexOfObject: [self cellAtRow:row column:column]];
+			NSInteger start = [[self cells] indexOfObject: [[self selectedCells] objectAtIndex: 0]];
+			NSInteger end = [[self cells] indexOfObject: [self cellAtRow:row column:column]];
 			
 			if( [[self selectedCells] containsObject:[self cellAtRow:row column:column]])
 				[self setSelectionFrom:end to:end anchor:end highlight: NO];
